+++ src/main/java/listeners/ItestListeners.java
@@ -1,12 +1,8 @@
 package listeners;
 
-import baseSetUp.BrowserInitialization;
-import org.openqa.selenium.WebDriver;
 import org.testng.ITestContext;
 import org.testng.ITestListener;
 import org.testng.ITestResult;
-
-import java.io.IOException;
 
 public class ItestListeners implements ITestListener
 {
