+++ src/main/java/seleniumUtils/SeleniumUtil.java
@@ -1,6 +1,5 @@
 package seleniumUtils;
 
-import baseSetUp.BrowserInitialization;
 import org.openqa.selenium.*;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.Select;
