+++ src/main/resources/Config.properties
@@ -1,6 +1,6 @@
 url = http://ecom-optimus.myshopify.com
 browser = chrome
-chromedriverpath = //Users//abhishek//Downloads//ui-automation//chromedriver
-firfoxdriverpath = //Users//abhishek//Downloads//ui-automation//geckodriver
+chromedriverpath = //Users//abhishek//Downloads//ui-automation//drivers//chromedriver
+firfoxdriverpath = //Users//abhishek//Downloads//ui-automation//drivers//geckodriver
 FailedTestScreenShort =  //Users//abhishek//Downloads//ui-automation//screenshots//FailureScreenshots//
 ActionScreenShots = /Users/abhishek/Downloads/ui-automation/screenshots/ActionScreenshots//
